requirements.txt
Django==3.2.9
django-environ==0.8.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
mysqlclient==2.1.0